- Replaced instances of NewParameters() with NewExtension() in session_test.go to ensure consistency in client extension handling.
- Updated SubscribeID and TrackPriority types to be based on uint64 and byte respectively, simplifying their definitions.
- Removed unnecessary tests for SubscribeID string representation, focusing on type validation.
- Adjusted TrackWriter to encode SubscribeID and GroupSequence as uint64.
- Enhanced version management by defining development and draft versions with clear constants.

+++ moqt/track_writer.go
@@ -133,8 +133,8 @@ func (s *TrackWriter) OpenGroup(seq GroupSequence) (*GroupWriter, error) {
 	}
 
 	err = message.GroupMessage{
-		SubscribeID:   s.subscribeID,
-		GroupSequence: seq,
+		SubscribeID:   uint64(s.subscribeID),
+		GroupSequence: uint64(seq),
 	}.Encode(stream)
 	if err != nil {
 		var strErr *quic.StreamError
